refactor: Move files into new cdc feature lib structure.
