+++ integration-libs/cdc/core/auth/facade/index.ts
