+++ integration-libs/cdc/public_api.ts
@@ -1,8 +1,8 @@
 /*
  * Public API Surface for CDC
  */
-export * from './src/auth/facade/index';
-export * from './src/auth/store/actions/index';
+export * from './core/auth/facade/index';
+export * from './core/auth/store/actions/index';
 export * from './src/cdc.module';
 export * from './components/gigya-raas/index';
 export * from './root/config/index';
