+++ integration-libs/cdc/root/cdc-root.module.ts
@@ -1,8 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { APP_INITIALIZER, ModuleWithProviders, NgModule } from '@angular/core';
 import { ConfigInitializerService, provideConfig } from '@spartacus/core';
+import { LogoutGuard } from '@spartacus/storefront';
 import { CdcConfig } from './config/cdc-config';
 import { CdcJsService } from './facade/cdc-js.service';
+import { CdcLogoutGuard } from './guards/cdc-logout.guard';
 
 export function cdcJsFactory(
   cdcJsService: CdcJsService,
@@ -17,6 +19,7 @@ export function cdcJsFactory(
 
 @NgModule({
   declarations: [],
+  providers: [{ provide: LogoutGuard, useExisting: CdcLogoutGuard }],
   imports: [CommonModule],
 })
 export class CdcRootModule {
